test(editor): add simple checks for fare_rules
