+++ src/test/java/com/conveyal/gtfs/util/FareDTO.java
@@ -17,13 +17,4 @@ public class FareDTO {
     public String agency_id;
     public Double transfer_duration;
     public FareRuleDTO[] fare_rules;
-
-    public class FareRuleDTO {
-        public int id;
-        public String fare_id;
-        public String route_id;
-        public String contains_id;
-        public String origin_id;
-        public String destination_id;
-    }
 }
