+++ src/test/java/com/conveyal/gtfs/util/FareRuleDTO.java
@@ -1,0 +1,10 @@
+package com.conveyal.gtfs.util;
+
+public class FareRuleDTO {
+    public int id;
+    public String fare_id;
+    public String route_id;
+    public String contains_id;
+    public String origin_id;
+    public String destination_id;
+}
